LayoutTests/http/tests/xmlhttprequest/access-control-preflight-sync-header-denied-expected.txt

CONSOLE MESSAGE: line 35: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?test=access-control-preflight-sync-header-denied&state=header. Request header field X-NON-STANDARD is not allowed by Access-Control-Allow-Headers.
PASS: Request successfully blocked.
